fix(android-sdk): SPEC-401-A R16 batch B — 4 parity gaps closed
Lens A P2 — Pricing card unselected background:
iOS ContentBlockStandaloneViews.swift:1902 uses Color(.systemBackground)
(opaque per theme); Android hardcoded Color.Transparent. With a tinted page
background the Android plan cards visibly lacked a solid backplate. Now
uses MaterialTheme.colorScheme.surface which adopts the right opaque color
in light/dark themes.

Lens A P2 — Pricing card highlighted shadow missing:
iOS ContentBlockStandaloneViews.swift:1911 adds `.shadow(color:
isHighlighted ? accent.opacity(0.15) : .clear, radius: 4, y: 2)`. Android
Card had zero shadow because we override containerColor (Material default
elevation goes to 0). Highlighted "best value" plan now lifts with the
accent-tinted shadow.

Lens A P3 — Timeline circle/icon/dot dimensions + connector + fontWeight:
iOS uses 28pt circle / 12pt checkmark / 10pt current dot / 20|32pt
connector; Android was 24/14/8 + 12|24 — visibly tighter side-by-side.
Also iOS keeps title fontWeight=.semibold for ALL items (only color varies
by status); Android was flipping to Normal for non-current items. Now
matches iOS dimensions + uniform SemiBold weight.

Lens B P3 — chat_completed safety-net duration_ms type:
ChatStepComposable.kt:617 emitted Long (missing .toInt()) only on the
LaunchedEffect safety-net path; sibling sites at :303 + :317 emit Int.
iOS at ChatStepView.swift:589 always Int. Schema drift across 3 chat_
completed sites resolved.

Deferred P3 — pulsing star background coordinate space drift (Android uses
abstract 0..1000 vs iOS UIScreen.bounds). Cosmetic only on non-fullscreen
blocks; not fixed this round.

Build: clean on mac-michal (23.8s, 0 errors, 22 warnings).

+++ src/main/kotlin/ai/appdna/sdk/onboarding/ChatStepComposable.kt
@@ -614,7 +614,12 @@ fun ChatStepComposable(
                         "completion_reason" to "max_turns",
                         "user_turn_count" to userTurnCount,
                         "total_messages" to messages.size,
-                        "duration_ms" to (System.currentTimeMillis() - startTime),
+                        // SPEC-401-A R16 — `.toInt()` to match iOS Int payload
+                        // at ChatStepView.swift:589 + the regular completion
+                        // paths (lines 303 + 317). Was emitting Long here only
+                        // on this safety-net path → schema drift across the
+                        // three chat_completed call sites.
+                        "duration_ms" to (System.currentTimeMillis() - startTime).toInt(),
                     ))
                 }
             }
